update: move mpn to wd_mpnn.py
