+++ ppmat/models/wd_mpnn/__init__.py
@@ -1,3 +1,3 @@
-from ppmat.models.wd_mpnn.model import WDMPNN
+from ppmat.models.wd_mpnn.wd_mpnn import WDMPNN
 
 __all__ = ["WDMPNN"]
